feat(app-generic): create filters service
