+++ apps/api/src/app/shared/constants.ts
@@ -1,2 +1,1 @@
 export const DAL_SERVICE = 'DalService';
-export const EXCEPTION_MESSAGE_ON_WEBHOOK_FILTER = 'Exception while performing webhook request.';
