+++ apps/api/src/app/shared/helpers/hmac.service.ts
@@ -1,6 +1,7 @@
 import { Logger } from '@nestjs/common';
 import { createHmac } from 'crypto';
 
+// TODO: Remove this and its usage rely on @novu/application-generic one
 export function createHash(key: string, valueToHash: string) {
   Logger.verbose('Creating Hmac');
