+++ apps/api/src/app/shared/shared.module.ts
@@ -32,6 +32,7 @@ import {
   DalServiceHealthIndicator,
   distributedLockService,
   featureFlagsService,
+  FilterConditionsService,
   getIsTopicNotificationEnabled,
   InvalidateCacheService,
   LoggerModule,
@@ -84,6 +85,7 @@ const PROVIDERS = [
   DalServiceHealthIndicator,
   distributedLockService,
   featureFlagsService,
+  FilterConditionsService,
   getIsTopicNotificationEnabled,
   InvalidateCacheService,
   storageService,
