+++ apps/worker/src/app/shared/shared.module.ts
@@ -40,6 +40,7 @@ import {
   distributedLockService,
   EventsDistributedLockService,
   featureFlagsService,
+  FilterConditionsService,
   InvalidateCacheService,
   LoggerModule,
   ProcessSubscriber,
@@ -109,6 +110,7 @@ const PROVIDERS = [
   distributedLockService,
   EventsDistributedLockService,
   featureFlagsService,
+  FilterConditionsService,
   getUseMergedDigestId,
   InvalidateCacheService,
   ProcessSubscriber,
