--- apps/worker/src/app/shared/utils/constants.ts
@@ -1,1 +1,0 @@
-export const EXCEPTION_MESSAGE_ON_WEBHOOK_FILTER = 'Exception while performing webhook request.';
