--- apps/worker/src/app/shared/utils/exceptions.ts
@@ -1,1 +1,0 @@
-export class PlatformException extends Error {}
