--- apps/worker/src/app/workflow/usecases/message-matcher/index.ts
@@ -1,1 +1,0 @@
-export { MessageMatcher } from './message-matcher.usecase';
