+++ apps/worker/src/app/workflow/usecases/send-message/digest/get-digest-events.usecase.ts
@@ -14,9 +14,8 @@ import {
   CreateExecutionDetails,
   CreateExecutionDetailsCommand,
   Instrument,
+  PlatformException,
 } from '@novu/application-generic';
-
-import { PlatformException } from '../../../../shared/utils';
 
 const LOG_CONTEXT = 'GetDigestEvents';
