+++ apps/worker/src/app/workflow/usecases/send-message/send-message-sms.usecase.ts
@@ -20,12 +20,12 @@ import {
   CompileTemplateCommand,
   SmsFactory,
   GetNovuProviderCredentials,
+  PlatformException,
 } from '@novu/application-generic';
 
 import { CreateLog } from '../../../shared/logs';
 import { SendMessageCommand } from './send-message.command';
 import { SendMessageBase } from './send-message.base';
-import { PlatformException } from '../../../shared/utils';
 
 @Injectable()
 export class SendMessageSms extends SendMessageBase {
