+++ packages/application-generic/package.json
@@ -109,7 +109,9 @@
     "@novu/zulip": "^0.21.0",
     "@sentry/node": "^7.12.1",
     "analytics-node": "^6.2.0",
+    "axios": "^1.3.3",
     "bullmq": "^3.10.2",
+    "chai": "^4.2.0",
     "class-transformer": "^0.5.1",
     "class-validator": "^0.14.0",
     "date-fns": "^2.29.2",
@@ -134,6 +136,7 @@
   "devDependencies": {
     "@istanbuljs/nyc-config-typescript": "^1.0.1",
     "@types/analytics-node": "^3.1.9",
+    "@types/chai": "^4.2.11",
     "@types/jest": "29.5.2",
     "@types/sinon": "^9.0.0",
     "codecov": "^3.5.0",
