+++ packages/application-generic/src/encryption/index.ts
@@ -1,1 +1,2 @@
 export * from './encrypt-provider';
+export { createHash } from './hmac';
