+++ packages/application-generic/src/index.ts
@@ -10,8 +10,6 @@ export * from './modules';
 export * from './usecases';
 export * from './instrumentation/index';
 export * from './utils/subscriber';
-export * from './utils/filter';
-export * from './utils/filter-processing-details';
 export * from './resilience';
 export * from './utils/exceptions';
 export * from './utils/email-normalization';
