+++ packages/application-generic/src/services/filters/index.ts
@@ -1,0 +1,6 @@
+export * from './filter-conditions.service';
+export {
+  FilterProcessingDetails,
+  IFilterData,
+  IFilterVariables,
+} from './filters.service';
