+++ packages/application-generic/src/services/index.ts
@@ -1,6 +1,14 @@
 export * from './in-memory-provider';
 export * from './launch-darkly.service';
 export * from './feature-flags.service';
+export {
+  FilterConditionsService,
+  FilterProcessingDetails,
+  IFilterConditionsResponse,
+  IFilterData,
+  IFilterVariables,
+  IUsedFilters,
+} from './filters';
 export * from './cache';
 export * from './queues';
 export * from './workers';
