--- packages/application-generic/src/usecases/conditions-filter/index.ts
@@ -1,2 +1,0 @@
-export * from './conditions-filter.command';
-export * from './conditions-filter.usecase';
